Initial commit of allocator and memory registration code
This code has been used in HPX for the libfabric parcelport
and experimentally in GHEX transport layer. It contains
allocators, memory pools and a memory pinning abstraction
with an implementation using libfabric.

There are limitations with the memory pool as it does not
release memory. This needs to be improved.